+++ CHANGELOG.md
@@ -188,3 +188,5 @@ FEATURES:
 ## v0.12.0
 
 - **Attribute Renamed:** The `technical_emails` attribute of a project resource is renamed to `technical_email`. It is now optional.
+- The `forward_rules` attribute of a load balancer resource is now validated with stricter rules to avoid errors.
+- Fixed some typos in the load balancer documentation.